requirements.txt
PySide6==6.10.2
pillow==12.1.0
opencv-python==4.11.0.86
numpy==2.2.4